refactor: move build_filter_condition to filters.py
