+++ apps/experiments/views/experiment.py
@@ -56,12 +56,7 @@ from apps.experiments.decorators import (
 )
 from apps.experiments.email import send_chat_link_email, send_experiment_invitation
 from apps.experiments.exceptions import ChannelAlreadyUtilizedException
-from apps.experiments.filters import (
-    build_participant_filter,
-    build_tags_filter,
-    build_timestamp_filter,
-    build_versions_filter,
-)
+from apps.experiments.filters import build_filter_condition
 from apps.experiments.forms import (
     ConsentForm,
     ExperimentForm,
@@ -184,7 +179,7 @@ class ExperimentSessionsTableView(SingleTableView, PermissionRequiredMixin):
             if not all([filter_column, filter_operator, filter_value]):
                 break
 
-            condition = self.build_filter_condition(filter_column, filter_operator, filter_value)
+            condition = build_filter_condition(filter_column, filter_operator, filter_value)
             if condition:
                 filter_conditions &= condition
                 filter_applied = True
@@ -192,20 +187,6 @@ class ExperimentSessionsTableView(SingleTableView, PermissionRequiredMixin):
         if filter_applied:
             query_set = query_set.filter(filter_conditions).distinct()
         return query_set
-
-    def build_filter_condition(self, column, operator, value):
-        """Build a Q object for the filter condition based on column, operator and value"""
-        if not value:
-            return None
-        if column == "participant":
-            return build_participant_filter(operator, value)
-        elif column == "last_message":
-            return build_timestamp_filter(operator, value)
-        elif column == "tags":
-            return build_tags_filter(operator, value)
-        elif column == "versions":
-            return build_versions_filter(operator, value)
-        return None
 
 
 class ExperimentVersionsTableView(SingleTableView, PermissionRequiredMixin):
